// frontend/src/generated/model/eventPlace.ts
/**
 * backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventPlace { 
    /**
     * A class representation of the BSON ObjectId type.
     */
    _id?: string;
    name: string;
    description: string;
    longitude: number;
    latitude: number;
    start_time: string;
    finish_time: string;
}